refactor: create v2 tests in test folder, move old tests to old folder
