+++ old/v2/docker-compose-2.yml
@@ -1,0 +1,11 @@
+version: '2'
+
+services:
+  web:
+    image: nginx:1.16.0
+    container_name: compose_test_web_2
+    command: 'nginx -g "daemon off;"'
+    environment:
+      NGINX_PORT: 8888
+    ports:
+      - 8888:8888
