+++ old/v2/docker-compose-42.yml
@@ -1,0 +1,12 @@
+version: '3'
+
+services:
+  some-service:
+    image: nginx:1.19.9-alpine
+    command: 'nginx -g "daemon off;"'
+    volumes:
+      # - ./volume:/mountedvolume
+      - db-data:/mountedsecondvolume
+
+volumes:
+  db-data:
