+++ old/v2/docker-compose.yml
@@ -1,0 +1,17 @@
+version: '2'
+
+services:
+  web:
+    image: nginx:1.16.0
+    container_name: compose_test_web
+    command: 'nginx -g "daemon off;"'
+    ports: 
+      - '0.0.0.0:80:80'
+      - '0.0.0.0:443:443'
+  proxy:
+    image: nginx:1.19.9-alpine
+    container_name: compose_test_proxy
+    command: 'nginx -g "daemon off;"'
+  hello:
+    image: hello-world
+    container_name: compose_test_hello
